fix minor bugs

+++ UserApp/app/src/main/java/com/hackathon/csec/foodie/Adapter/MenusAdapter.java
@@ -57,7 +57,7 @@ public class MenusAdapter extends RecyclerView.Adapter<MenusAdapter.ViewHolder> 
                 distance = distance + "away";
 
                 holder.distanceRes.setText(distance);
-                holder.ratingRes.setRating(arrayList.get(position).getRating());
+                holder.ratingRes.setRating((float) arrayList.get(position).getRating());
             }
 
             Glide.with(context).load(arrayList.get(position).getPhoto()).asBitmap().
